requirements.txt
Flask
openai
google-api-python-client
matplotlib
seaborn
pandas
stripe
python-dotenv
